Fixes infinite looping (#7)
Limits attempts at temp-file creation to avoid infinite loops, returning
the most recent `'EEXIST'` exception if a suitable filename cannot be
found.

Note that the number of retries attempted is equal to the total number
of filenames possible for a given template. This means that failure may
(and for n > 1 filenames likely _will_) occur before all possible
filenames are exhausted. This trade-off is intended to provide a
"reasonable" number of retry attempts without requiring a full scan/log
of previous attempts.